Fix clang-tidy warnings in 12 additional production files
- Fixed null pointer dereference warnings with proper null checks
- Replaced insecure strcat() calls with safe_concat()
- Fixed implicit type conversions with explicit casts
- Fixed unused parameter warnings
- Fixed sign conversion warnings
- Fixed narrowing conversion errors
- Fixed undeclared identifier errors
- Fixed override keyword issues
- Fixed read-only variable assignment warnings
- Implemented missing virtual functions in RemotePrinter

Files fixed:
- main/business/check.cc
- main/hardware/remote_printer.cc
- main/hardware/terminal.cc
- src/core/generic_char.cc
- src/network/vt_ccq_pipe.cc
- term/layer.cc
- term/term_credit_cheq.cc
- term/term_view.cc
- zone/button_zone.cc
- zone/order_zone.cc
- zone/pos_zone.cc
- zone/zone.cc

+++ term/layer.cc
@@ -1072,9 +1072,9 @@ int Layer::Triangle(int tx, int ty, int tw, int th, int image)
     short far_y = ty + (th - 1);
 
     XPoint pts[] = {
-        {mid_x, ty},        // Top
-        {tx,    far_y},     // Bottom-left
-        {far_x, far_y}      // Bottom-right
+        {static_cast<short>(mid_x), static_cast<short>(ty)},        // Top
+        {static_cast<short>(tx),    static_cast<short>(far_y)},     // Bottom-left
+        {static_cast<short>(far_x), static_cast<short>(far_y)}      // Bottom-right
     };
 
     XSetTSOrigin(dis, gfx, page_x, page_y);
